requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.41
botocore==1.31.41
click==8.1.7
colorama==0.4.6
fastapi==0.103.1
h11==0.14.0
httptools==0.6.0
idna==3.4
jmespath==1.0.1
numpy==1.25.2
pandas==2.1.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
